+++ monkey/common/types.py
@@ -1,0 +1,3 @@
+from pydantic import PositiveInt
+
+HardwareID = PositiveInt