+++ monkey/monkey_island/cc/resources/__init__.py
@@ -9,3 +9,4 @@ from .agent_configuration import AgentConfiguration
 from .pba_file_upload import PBAFileUpload, LINUX_PBA_TYPE, WINDOWS_PBA_TYPE
 from .pba_file_download import PBAFileDownload
 from .events import Events
+from .agents import Agents